Initial commit of lgpgame.
lgpgame is the working title for the game program that I am working on
to test out and improve libgamepieces.  The intent at this time is to
create a randomly generated maze that the player will try to find their
way out of.

Initially the game itself will be somewhat simplistic, but it will
improve and expand with time.  The maze will be 3D in nature though, so
it will still be rather large from the start.