Flesshed out class game::Maze.
I fleshed out class game::Maze, but it is still acting a bit odd.  It is
generating half as many rooms as I expected.  I need to look at the
algorythm a bit.

The two points that are of the most interest at the moment are
Maze::generateMaze() and Maze::pickDirection();