Added a Player class and generator.  Fixed the Maze Gen code.
I have created a class game::Player : public lgp::Creature, and class
game::GenPlayer : public lgp::IGenerator.  I decided that since I know
that the player object will end up getting considerable amounts of
project specific code associated with it, it is easier to change it to
its own class now, rather than let it get burried in the code as a
lgp::Creature first.

I noticed earlier that I was only getting around 650 or so objects after
running maze generation code.  This isn't exactly what I had in mind for
the size of the maze, given that it is an 11x11x11 cube.  However, after
failing to find the bug, I moved on to adding code to handle placement
of the start and finish rooms.  I discovered that the program was 100%
of the time finding a place to put the finish line, but also 100% of the
time failing to find a proper place to put the starting point.

I realized that I was using  0..10 as the range for coordinates in the
room placement code.  The problem is that the code that actually places
the room into the array mapM, already maps the world coordinates to the
appropriate numbers for the array.  (i.e. -5 becomes 0 automatically).
This means that it was not apropriate to manipulate the numbers the way
that I was in the room creation code.  The end result was that all the
rooms were being placed in the front, top half of the map.

This code has been changed to look for coordinates in a range of -5..5
now, and is functioning as it should (as far as I can tell at least).

I am ready now to write the code to put the player in the starting room,
and to start on the code to handle text commands and movement. Wish me
luck. :-)